// src/desafios/interfaces/desafio.interface.ts
import { Document } from "mongoose";
import { Jogador } from "src/jogadores/interfaces/jogador.interface";
import { Partida } from "src/partidas/interfaces/partidas.interface";

// Generated by https://quicktype.io

export interface Desafio extends Document {
    dataHoraDesafio:     Date;
    status:              DesafioStatus;
    dataHoraSolicitacao: Date;
    dataHoraResposta:    Date;
    solicitante:         Jogador;
    categoria:           string;
    jogadores:           Jogador[];
    partida:             Partida;
}

export enum DesafioStatus {
    REALIZADO = 'REALIZADO',
    PENDENTE = 'PENDENTE',
    ACEITO = 'ACEITO',
    NEGADO = 'NEGADO',
    CANCELADO = 'CANCELADO'
}